+++ app/Http/Controllers/CommentController.php
@@ -61,6 +61,9 @@ class CommentController extends Controller
     {
         $id = $request->route('id');
         $comment = Comment::find($id);
+        if (!$comment) {
+            return "not Author";
+        }
         $comment->content = $request->input('content');
         $comment->status = $request->input('status');
         $comment->author_id = $request->input('author_id');
@@ -76,6 +79,9 @@ class CommentController extends Controller
     {
         $id = $request->route('id');
         $commentId = Comment::find($id);
+        if (!$commentId) {
+            return "not Author";
+        }
         return response()->json($commentId, 200);
     }
 
@@ -83,6 +89,9 @@ class CommentController extends Controller
     {
         $id = $request->route('id');
         $comment = Comment::find($id);
+        if (!$comment) {
+            return "not Author";
+        }
         $comment->delete();
         Log::info("delete success");
         return "success";
